requirements.txt
slackclient==1.3.1
python-decouple==3.1
sqlalchemy==1.3.1